requirements.txt
streamlit
pandas
python-docx
google-generativeai>=0.5.0

